pyproject.toml
[tool.poetry]
name = "mandacaru-dev-ibiapaba-cdd"
version = "0.1.0"
description = "Projeto do desafio \"Análise de Sentimentos\" do bootcamp Mandacaru.dev - módulo Ibiapaba"
authors = ["Elvin Matheus <elvinmatheussousa@gmail.com>"]
readme = "README.md"
packages = [{include = "mandacaru"}]

[tool.poetry.dependencies]
python = "^3.8.10"
numpy = "^1.22.3"
pandas = "^1.4.2"
matplotlib = "^3.5.1"
jupyterlab = "^3.4.4"
seaborn = "^0.11.2"
scikit-learn = "^1.3.0"
fastapi = "^0.104.1"
kaggle = "^1.5.16"
pdoc3 = "^0.10.0"
hydra-core = "^1.3.2"
wordcloud = "^1.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
